Added Syncronization code. Added code to set files to readonly when rights are readonly. Fixed delete.
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@2633 db2431b9-df06-0410-971c-a38ea9a75485

+++ simias/src/Sync/SyncFile.cs
@@ -336,7 +336,9 @@ namespace Simias.Sync
 				}
 			}
 			// We need to delete the temp file.
-			File.Delete(workFile);
+
+			if (workFile != null)
+				File.Delete(workFile);
 		}
 
 		#endregion
